Refactored mobile slashdot url parsing and title fetching into separate classes
