+++ src/main/java/com/gorlah/kappabot/function/slashdot/MobileSlashdotTitleParser.java
@@ -1,0 +1,35 @@
+package com.gorlah.kappabot.function.slashdot;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Component;
+
+import java.util.List;
+import java.util.Objects;
+import java.util.regex.MatchResult;
+import java.util.regex.Pattern;
+import java.util.stream.Collectors;
+
+@Slf4j
+@Component
+@RequiredArgsConstructor
+public class MobileSlashdotTitleParser {
+
+    private final Pattern MOBILE_SLASHDOT_ARTICLE_URL_PATTERN
+            = Pattern.compile("https://m\\.slashdot\\.org/story/\\d+");
+
+    private final MobileSlashdotTitleFetcher mobileSlashdotTitleFetcher;
+
+    public boolean containsMobileSlashdotUrl(String message) {
+        return MOBILE_SLASHDOT_ARTICLE_URL_PATTERN.matcher(message).find();
+    }
+
+    public List<String> getTitles(String message) {
+        return MOBILE_SLASHDOT_ARTICLE_URL_PATTERN.matcher(message).results()
+                .parallel()
+                .map(MatchResult::group)
+                .map(mobileSlashdotTitleFetcher::getTitle)
+                .filter(Objects::nonNull)
+                .collect(Collectors.toUnmodifiableList());
+    }
+}
