+++ src/main/resources/application.properties
@@ -10,6 +10,11 @@ discord.bot.status=Hello there \uD83E\uDD86
 # Imgur
 imgur.clientId=
 
+# Slashdot
+slashdot.titleFetcher.userAgent=Mozilla/5.0 (Windows NT 10.0; Win64; x64) \
+  AppleWebKit/537.36 (KHTML, like Gecko) \
+  Chrome/77.0.3865.120 Safari/537.36
+
 # OpenWeather
 # https://home.openweathermap.org/api_keys
 openweather.apiKey=
